Work in progress. Most of ordermatch code refactored to Rust.

+++ mm2src/coins/lp_coins.rs
@@ -93,14 +93,14 @@ pub type TransactionFut = Box<dyn Future<Item=TransactionEnum, Error=String>>;
 
 /// Swap operations (mostly based on the Hash/Time locked transactions implemented by coin wallets).
 pub trait SwapOps {
-    fn send_taker_fee(&self, fee_addr: &[u8], amount: u64) -> TransactionFut;
+    fn send_taker_fee(&self, fee_addr: &[u8], amount: BigDecimal) -> TransactionFut;
 
     fn send_maker_payment(
         &self,
         time_lock: u32,
         taker_pub: &[u8],
         secret_hash: &[u8],
-        amount: u64
+        amount: BigDecimal,
     ) -> TransactionFut;
 
     fn send_taker_payment(
@@ -108,7 +108,7 @@ pub trait SwapOps {
         time_lock: u32,
         maker_pub: &[u8],
         secret_hash: &[u8],
-        amount: u64,
+        amount: BigDecimal,
     ) -> TransactionFut;
 
     fn send_maker_spends_taker_payment(
@@ -147,7 +147,7 @@ pub trait SwapOps {
         &self,
         fee_tx: TransactionEnum,
         fee_addr: &[u8],
-        amount: u64
+        amount: BigDecimal,
     ) -> Result<(), String>;
 
     fn validate_maker_payment(
@@ -156,7 +156,7 @@ pub trait SwapOps {
         time_lock: u32,
         maker_pub: &[u8],
         priv_bn_hash: &[u8],
-        amount: u64,
+        amount: BigDecimal,
     ) -> Result<(), String>;
 
     fn validate_taker_payment(
@@ -165,7 +165,7 @@ pub trait SwapOps {
         time_lock: u32,
         taker_pub: &[u8],
         priv_bn_hash: &[u8],
-        amount: u64,
+        amount: BigDecimal,
     ) -> Result<(), String>;
 }
 
@@ -212,7 +212,7 @@ struct WithdrawRequest {
     coin: String,
     to: String,
     #[serde(default)]
-    amount: f64,
+    amount: BigDecimal,
     #[serde(default)]
     max: bool
 }
@@ -265,7 +265,7 @@ pub trait MmCoin: SwapOps + MarketCoinOps + IguanaInfo + Debug + 'static {
 
     fn can_i_spend_other_payment(&self) -> Box<Future<Item=(), Error=String> + Send>;
 
-    fn withdraw(&self, to: &str, amount: f64, max: bool) -> Box<Future<Item=TransactionDetails, Error=String> + Send>;
+    fn withdraw(&self, to: &str, amount: BigDecimal, max: bool) -> Box<Future<Item=TransactionDetails, Error=String> + Send>;
 
     /// Maximum number of digits after decimal point used to denominate integer coin units (satoshis, wei, etc.)
     fn decimals(&self) -> u8;
